requirements.txt
django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
requests>=2.31
pytest>=7.4
pytest-django>=4.5